added and set up alembic migrations and added initial revision

+++ backend/app/models/simulation_params.py
@@ -1,11 +1,11 @@
-from backend.app.database import Base
-from sqlalchemy import Column, String, Integer, func, TIMESTAMP, Float
+from .base import Base
+from sqlalchemy import Column, String, SmallInteger, func, TIMESTAMP, Float
 
 
 class SimulationParams(Base):
     __tablename__ = "simulation_params"
 
-    id = Column(Integer, primary_key=True, nullable=False)
+    id = Column(SmallInteger, primary_key=True, nullable=False)
     wind_speed = Column(Float, nullable=False)
     wind_direction = Column(Float, nullable=False) # Тут будет измеряться вградусах
     stability_class = Column(String, nullable=False)
